+++ modal/config.py
@@ -150,9 +150,6 @@ def _check_config() -> None:
         # But we want to give users time to activate one of their profiles without disruption
         message = dedent(
             """
-
-            WARNING:
-
             Support for using an implicit 'default' profile is deprecated.
             Please use `modal profile activate` to activate one of your profiles.
             (Use `modal profile list` to see the options.)
@@ -160,7 +157,7 @@ def _check_config() -> None:
             This will become an error in a future update.
             """
         )
-        deprecation_warning(date(2024, 2, 6), message)
+        deprecation_warning(date(2024, 2, 6), message, show_source=False)
 
 
 if "MODAL_ENV" in os.environ:
@@ -197,7 +194,7 @@ _SETTINGS = {
     "environment": _Setting(),
     "default_cloud": _Setting(None, transform=lambda x: x if x else None),
     "worker_id": _Setting(),  # For internal debugging use.
-    "restore_state_path": _Setting("/opt/modal/restore-state.json"),
+    "restore_state_path": _Setting("/__modal/restore-state.json"),
 }