+++ modal/dict.py
@@ -99,7 +99,7 @@ class _Dict(_Object, type_prefix="di"):
             req = api_pb2.DictGetOrCreateRequest(
                 deployment_name=label,
                 namespace=namespace,
-                environment_name=_get_environment_name(environment_name),
+                environment_name=_get_environment_name(environment_name, resolver),
                 object_creation_type=(api_pb2.OBJECT_CREATION_TYPE_CREATE_IF_MISSING if create_if_missing else None),
             )
             response = await resolver.client.stub.DictGetOrCreate(req)