+++ modal/image.py
@@ -1052,11 +1052,13 @@ class _Image(_Object, type_prefix="im"):
         add_python: Optional[str] = None,
         **kwargs,
     ) -> "_Image":
-        """Build a Modal image from a private image in GCP Artifact Registry.
+        """Build a Modal image from a private image in Google Cloud Platform (GCP) Artifact Registry.
 
-        You will need to pass a `modal.Secret` containing your GCP service account key
+        You will need to pass a `modal.Secret` containing [your GCP service account key data](https://cloud.google.com/iam/docs/keys-create-delete#creating)
         as `SERVICE_ACCOUNT_JSON`. This can be done from the [Secrets](/secrets) page.
-        The service account needs to have at least an "Artifact Registry Reader" role.
+        The service account needs to have at least an ["Artifact Registry Reader"](https://cloud.google.com/artifact-registry/docs/access-control#roles) role.
+
+        **Note:** This method does not use `GOOGLE_APPLICATION_CREDENTIALS` as that variable accepts a path to a JSON file, not the actual JSON string.
 
         See `Image.from_registry()` for information about the other parameters.
 
@@ -1065,11 +1067,13 @@ class _Image(_Object, type_prefix="im"):
         ```python
         modal.Image.from_gcp_artifact_registry(
             "us-east1-docker.pkg.dev/my-project-1234/my-repo/my-image:my-version",
-            secrets=[modal.Secret.from_name("my-gcp-secret")],
+            secret=modal.Secret.from_name("my-gcp-secret"),
             add_python="3.11",
         )
         ```
         """
+        if "secrets" in kwargs:
+            raise TypeError("Passing a list of 'secrets' is not supported; use the singular 'secret' argument.")
         image_registry_config = _ImageRegistryConfig(api_pb2.REGISTRY_AUTH_TYPE_GCP, secret)
         return _Image.from_registry(
             tag,
@@ -1106,11 +1110,13 @@ class _Image(_Object, type_prefix="im"):
         ```python
         modal.Image.from_aws_ecr(
             "000000000000.dkr.ecr.us-east-1.amazonaws.com/my-private-registry:my-version",
-            secrets=[modal.Secret.from_name("aws")],
+            secret=modal.Secret.from_name("aws"),
             add_python="3.11",
         )
         ```
         """
+        if "secrets" in kwargs:
+            raise TypeError("Passing a list of 'secrets' is not supported; use the singular 'secret' argument.")
         image_registry_config = _ImageRegistryConfig(api_pb2.REGISTRY_AUTH_TYPE_AWS, secret)
         return _Image.from_registry(
             tag,
@@ -1361,6 +1367,18 @@ class _Image(_Object, type_prefix="im"):
 
     @contextlib.contextmanager
     def imports(self):
+        """
+        Used to import packages in global scope that are only available when running remotely.
+        By using this context manager you can avoid an `ImportError` due to not having certain
+        packages installed locally.
+
+        **Usage:**
+
+        ```python notest
+        with image.imports():
+            import torch
+        ```
+        """
         env_image_id = config.get("image_id")
         try:
             yield
@@ -1378,8 +1396,9 @@ class _Image(_Object, type_prefix="im"):
     def run_inside(self):
         """`Image.run_inside` is deprecated - use `Image.imports` instead.
 
-        Usage:
-        ```
+        **Usage:**
+
+        ```python notest
         with image.imports():
             import torch
         ```