+++ modal/stub.py
@@ -2,7 +2,6 @@
 import inspect
 import os
 import typing
-import warnings
 from datetime import date
 from pathlib import PurePosixPath
 from typing import Any, AsyncGenerator, Callable, ClassVar, Dict, List, Optional, Sequence, Tuple, Union
@@ -30,7 +29,6 @@ from .network_file_system import _NetworkFileSystem
 from .object import _Object
 from .partial_function import PartialFunction, _PartialFunction
 from .proxy import _Proxy
-from .queue import _Queue
 from .retries import Retries
 from .runner import _run_stub
 from .s3mount import _S3Mount
@@ -319,18 +317,6 @@ class _Stub:
         else:
             return _default_image
 
-    @property
-    def _pty_input_stream(self):
-        return self._indexed_objects.get("_pty_input_stream", None)
-
-    def _add_pty_input_stream(self):
-        if self._pty_input_stream:
-            warnings.warn(
-                "Running multiple interactive functions at the same time is not fully supported, and could lead to unexpected behavior."
-            )
-        else:
-            self._indexed_objects["_pty_input_stream"] = _Queue.new()
-
     def _get_watch_mounts(self):
         all_mounts = [
             *self._mounts,
@@ -534,9 +520,6 @@ class _Stub:
             if is_generator is None:
                 is_generator = inspect.isgeneratorfunction(raw_f) or inspect.isasyncgenfunction(raw_f)
 
-            if interactive:
-                self._add_pty_input_stream()
-
             function = _Function.from_args(
                 info,
                 stub=self,
@@ -603,6 +586,7 @@ class _Stub:
         checkpointing_enabled: bool = False,  # Enable memory checkpointing for faster cold starts.
         block_network: bool = False,  # Whether to block network access
         secret: Optional[_Secret] = None,  # Deprecated: use `secrets`
+        _allow_background_volume_commits: bool = False,
         max_inputs: Optional[
             int
         ] = None,  # Limits the number of inputs a container handles before shutting down. Use `max_inputs = 1` for single-use containers.
@@ -634,6 +618,7 @@ class _Stub:
             cloud=cloud,
             checkpointing_enabled=checkpointing_enabled,
             block_network=block_network,
+            _allow_background_volume_commits=_allow_background_volume_commits,
             max_inputs=max_inputs,
         )