+++ modal_test_support/functions.py
@@ -281,3 +281,25 @@ class BuildCls:
     @method()
     def f(self, x):
         return self._k * x
+
+
+@stub.cls(checkpointing_enabled=True)
+class CheckpointingCls:
+    def __init__(self):
+        self._vals = []
+
+    @enter(checkpoint=True)
+    def enter1(self):
+        self._vals.append("A")
+
+    @enter(checkpoint=True)
+    def enter2(self):
+        self._vals.append("B")
+
+    @enter()
+    def enter3(self):
+        self._vals.append("C")
+
+    @method()
+    def f(self, x):
+        return "".join(self._vals) + x